[fix] Fixes --usage as docs
* The objective was for --help to provide adequate
documentation for a project. However at this point we were
removing anything that was marked as sensitive, including
url, passwords, etc. This would not allow people to
understand what to pass to frameless in order to get
the desired output

This patch addresses that problem by showing the default
in the default behavior, and the config file only when it
is not sensitive